Refactoring Grids, Functions and RNG
Refactored the grid into a Color Grid, and a Dungeon, and moved useful functions and data structures to a module called system. Changed the Math.random to a random number generator that can be seeded, and a text input to allow me to input the seed if I want to, but it defaults to the current time.