Info UI, Magnify, and Bug Fixes
Added a bunch of textboxes that contain all the info used to generate the dungeon. Added magnification buttons that can magnify, the dungeon, and fixed the issue with the mouse dragging not working, and if the cells are showing then they are still showing after clicking one of the regenerate buttons.