IsConnected, Split off ColorGrid, and  Use Info from Textboxes
The info in the textboxes now updates the dungeon generation info, and it's used when you click on the Regenerate button. Added the IsConnected algorithm so that the Regenerate button always creates a connected dungeon, and took out ColorGrid into its own file structure since it doesn't have anything to do with the dungeon.